Aim and Fire stages work
Tank can switch stage between aiming and firing. The bullet has a link back to the tank and the explosion effect has also been linked to the bullet.
Next step: Investigate explosive force and calculate distance between the bullet and the things it collides against.
Make an enemy tank so we can start calculating damage.